+++ src/NuGetForUnity.Tests/Assets/Tests/Editor/NuGetTests.cs
@@ -7,8 +7,8 @@ using System.Reflection;
 using System.Threading.Tasks;
 using NugetForUnity;
 using NugetForUnity.Configuration;
-using NugetForUnity.Data;
 using NugetForUnity.Helper;
+using NugetForUnity.Models;
 using NugetForUnity.PackageSource;
 using NUnit.Framework;
 using UnityEditor;
