+++ src/NuGetForUnity/Editor/InstalledPackagesManager.cs
@@ -4,8 +4,8 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using NugetForUnity.Configuration;
-using NugetForUnity.Data;
 using NugetForUnity.Helper;
+using NugetForUnity.Models;
 using NugetForUnity.PackageSource;
 using Debug = UnityEngine.Debug;
