+++ src/NuGetForUnity/Editor/Models.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0f6275a4809a54344af845333ab7c816
+guid: 789926601f5c1da4593610a1e60dee73
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
