+++ src/NuGetForUnity/Editor/Models/INugetPackage.cs
@@ -3,7 +3,7 @@ using System.Threading.Tasks;
 using NugetForUnity.PackageSource;
 using UnityEngine;
 
-namespace NugetForUnity.Data
+namespace NugetForUnity.Models
 {
     /// <summary>
     ///     Represents a NuGet package.
