+++ src/NuGetForUnity/Editor/Models/INugetPackage.cs.meta
