+++ src/NuGetForUnity/Editor/Models/INugetPackageIdentifier.cs.meta
