+++ src/NuGetForUnity/Editor/Models/NugetFrameworkGroup.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace NugetForUnity.Data
+namespace NugetForUnity.Models
 {
     /// <summary>
     ///     Represents a grouping of items by framework type and version.
