+++ src/NuGetForUnity/Editor/Models/NugetFrameworkGroup.cs.meta
