+++ src/NuGetForUnity/Editor/Models/NugetPackageIdentifier.cs
@@ -2,7 +2,7 @@
 using System.Diagnostics.CodeAnalysis;
 using UnityEngine;
 
-namespace NugetForUnity.Data
+namespace NugetForUnity.Models
 {
     /// <summary>
     ///     Represents an identifier for a NuGet package. It contains only an ID and a Version number.
