+++ src/NuGetForUnity/Editor/Models/NugetPackageIdentifier.cs.meta
