+++ src/NuGetForUnity/Editor/Models/NugetPackageLocal.cs.meta
