+++ src/NuGetForUnity/Editor/Models/NugetPackageV2.cs.meta
