+++ src/NuGetForUnity/Editor/Models/NugetPackageV2Base.cs.meta
