+++ src/NuGetForUnity/Editor/Models/NugetPackageV3.cs.meta
