+++ src/NuGetForUnity/Editor/Models/NugetPackageVersion.cs.meta
