+++ src/NuGetForUnity/Editor/Models/SerializableNugetPackage.cs.meta
