+++ src/NuGetForUnity/Editor/PackageSource/INugetPackageSource.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using NugetForUnity.Data;
+using NugetForUnity.Models;
 
 namespace NugetForUnity.PackageSource
 {
