+++ src/NuGetForUnity/Editor/PackageSource/NugetApiClientV3.cs
@@ -10,8 +10,8 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using NugetForUnity.Data;
 using NugetForUnity.Helper;
+using NugetForUnity.Models;
 using UnityEditor;
 using UnityEngine;
