+++ src/NuGetForUnity/Editor/PackageSource/NugetPackageSourceCombined.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using NugetForUnity.Data;
+using NugetForUnity.Models;
 using UnityEngine;
 
 namespace NugetForUnity.PackageSource
