+++ src/NuGetForUnity/Editor/PackageSource/NugetPackageSourceLocal.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using NugetForUnity.Configuration;
-using NugetForUnity.Data;
+using NugetForUnity.Models;
 using UnityEngine;
 
 namespace NugetForUnity.PackageSource
