+++ src/NuGetForUnity/Editor/PackageSource/NugetPackageSourceV3.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using NugetForUnity.Data;
+using NugetForUnity.Models;
 using UnityEngine;
 
 namespace NugetForUnity.PackageSource
