+++ src/NuGetForUnity/Editor/UnityPreImportedLibraryResolver.cs
@@ -2,7 +2,7 @@
 using System.IO;
 using System.Linq;
 using NugetForUnity.Configuration;
-using NugetForUnity.Data;
+using NugetForUnity.Models;
 using UnityEditor;
 using UnityEditor.Compilation;
